+++ src/config_parser.c
@@ -72,96 +72,6 @@ typedef struct tokenptr {
 
 #include "GENERATED_config_tokens.h"
 
-/*
- * Pushes a string (identified by 'identifier') on the stack. We simply use a
- * single array, since the number of entries we have to store is very small.
- *
- */
-static void push_string(struct stack *ctx, const char *identifier, const char *str) {
-    for (int c = 0; c < 10; c++) {
-        if (ctx->stack[c].identifier != NULL &&
-            strcmp(ctx->stack[c].identifier, identifier) != 0) {
-            continue;
-        }
-        if (ctx->stack[c].identifier == NULL) {
-            /* Found a free slot, let’s store it here. */
-            ctx->stack[c].identifier = identifier;
-            ctx->stack[c].val.str = sstrdup(str);
-            ctx->stack[c].type = STACK_STR;
-        } else {
-            /* Append the value. */
-            char *prev = ctx->stack[c].val.str;
-            sasprintf(&(ctx->stack[c].val.str), "%s,%s", prev, str);
-            free(prev);
-        }
-        return;
-    }
-
-    /* When we arrive here, the stack is full. This should not happen and
-     * means there’s either a bug in this parser or the specification
-     * contains a command with more than 10 identified tokens. */
-    fprintf(stderr, "BUG: config_parser stack full. This means either a bug "
-                    "in the code, or a new command which contains more than "
-                    "10 identified tokens.\n");
-    exit(EXIT_FAILURE);
-}
-
-static void push_long(struct stack *ctx, const char *identifier, long num) {
-    for (int c = 0; c < 10; c++) {
-        if (ctx->stack[c].identifier != NULL) {
-            continue;
-        }
-        /* Found a free slot, let’s store it here. */
-        ctx->stack[c].identifier = identifier;
-        ctx->stack[c].val.num = num;
-        ctx->stack[c].type = STACK_LONG;
-        return;
-    }
-
-    /* When we arrive here, the stack is full. This should not happen and
-     * means there’s either a bug in this parser or the specification
-     * contains a command with more than 10 identified tokens. */
-    fprintf(stderr, "BUG: config_parser stack full. This means either a bug "
-                    "in the code, or a new command which contains more than "
-                    "10 identified tokens.\n");
-    exit(EXIT_FAILURE);
-}
-
-static const char *get_string(const struct stack *ctx, const char *identifier) {
-    for (int c = 0; c < 10; c++) {
-        if (ctx->stack[c].identifier == NULL) {
-            break;
-        }
-        if (strcmp(identifier, ctx->stack[c].identifier) == 0) {
-            return ctx->stack[c].val.str;
-        }
-    }
-    return NULL;
-}
-
-static long get_long(const struct stack *ctx, const char *identifier) {
-    for (int c = 0; c < 10; c++) {
-        if (ctx->stack[c].identifier == NULL) {
-            break;
-        }
-        if (strcmp(identifier, ctx->stack[c].identifier) == 0) {
-            return ctx->stack[c].val.num;
-        }
-    }
-    return 0;
-}
-
-static void clear_stack(struct stack *ctx) {
-    for (int c = 0; c < 10; c++) {
-        if (ctx->stack[c].type == STACK_STR) {
-            free(ctx->stack[c].val.str);
-        }
-        ctx->stack[c].identifier = NULL;
-        ctx->stack[c].val.str = NULL;
-        ctx->stack[c].val.num = 0;
-    }
-}
-
 /*******************************************************************************
  * The parser itself.
  ******************************************************************************/
